rewrote CAWR and REX, added support for a port via config.json
now both CAWR and REX are able to resume training so long as you have saved a state.

REX is no long just REX anymore, and is now RexAnnealingWarmRestarts, or RAWR for short as it now anneals with warm restarts, if you want it to behave like it used to, all you need to do is set the first cycle max steps to total steps, and for backup, probably should set gamma to 1.